LPS-53660 move code into capability
Call the capability from the servicess to keep the tests passing.
