+++ portal-impl/src/com/liferay/portlet/documentlibrary/service/impl/DLAppServiceImpl.java
@@ -25,6 +25,7 @@ import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.repository.InvalidRepositoryIdException;
 import com.liferay.portal.kernel.repository.Repository;
 import com.liferay.portal.kernel.repository.RepositoryException;
+import com.liferay.portal.kernel.repository.capabilities.ProcessorCapability;
 import com.liferay.portal.kernel.repository.capabilities.TrashCapability;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.repository.model.FileVersion;
@@ -46,6 +47,7 @@ import com.liferay.portal.kernel.util.TempFileEntryUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.model.Lock;
+import com.liferay.portal.repository.capabilities.LiferayProcessorCapability;
 import com.liferay.portal.security.permission.ActionKeys;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portlet.documentlibrary.NoSuchFileEntryException;
@@ -454,7 +456,10 @@ public class DLAppServiceImpl extends DLAppServiceBaseImpl {
 
 		FileEntry fileEntry = repository.getFileEntry(fileEntryId);
 
-		DLProcessorRegistryUtil.cleanUp(fileEntry.getLatestFileVersion());
+		ProcessorCapability processorCapability =
+			new LiferayProcessorCapability();
+
+		processorCapability.cleanUp(fileEntry.getLatestFileVersion());
 
 		FileVersion draftFileVersion = repository.cancelCheckOut(fileEntryId);
 
@@ -2912,7 +2917,10 @@ public class DLAppServiceImpl extends DLAppServiceBaseImpl {
 			getUserId(), fileEntryId, sourceFileName, mimeType, title,
 			description, changeLog, majorVersion, file, serviceContext);
 
-		DLProcessorRegistryUtil.cleanUp(fileEntry.getLatestFileVersion());
+		ProcessorCapability processorCapability =
+			new LiferayProcessorCapability();
+
+		processorCapability.cleanUp(fileEntry.getLatestFileVersion());
 
 		dlAppHelperLocalService.updateFileEntry(
 			getUserId(), fileEntry, null, fileEntry.getFileVersion(),
@@ -3004,7 +3012,10 @@ public class DLAppServiceImpl extends DLAppServiceBaseImpl {
 			description, changeLog, majorVersion, is, size, serviceContext);
 
 		if (is != null) {
-			DLProcessorRegistryUtil.cleanUp(fileEntry.getLatestFileVersion());
+			ProcessorCapability processorCapability =
+				new LiferayProcessorCapability();
+
+			processorCapability.cleanUp(fileEntry.getLatestFileVersion());
 
 			oldFileVersion = null;
 		}
@@ -3035,7 +3046,10 @@ public class DLAppServiceImpl extends DLAppServiceBaseImpl {
 			getUserId(), fileEntryId, sourceFileName, mimeType, title,
 			description, changeLog, majorVersion, file, serviceContext);
 
-		DLProcessorRegistryUtil.cleanUp(fileEntry.getLatestFileVersion());
+		ProcessorCapability processorCapability =
+			new LiferayProcessorCapability();
+
+		processorCapability.cleanUp(fileEntry.getLatestFileVersion());
 
 		repository.checkInFileEntry(
 			getUserId(), fileEntryId, majorVersion, changeLog, serviceContext);
@@ -3066,7 +3080,10 @@ public class DLAppServiceImpl extends DLAppServiceBaseImpl {
 			description, changeLog, majorVersion, is, size, serviceContext);
 
 		if (is != null) {
-			DLProcessorRegistryUtil.cleanUp(fileEntry.getLatestFileVersion());
+			ProcessorCapability processorCapability =
+				new LiferayProcessorCapability();
+
+			processorCapability.cleanUp(fileEntry.getLatestFileVersion());
 
 			oldFileVersion = null;
 		}
